+++ CHANGELOG.md
@@ -2,17 +2,25 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [0.3.5] - 2020-08-09
+### Changed
+- Fix bug in handling of nested classes in code generation
+
 ## [0.3.4] - 2020-08-09
 ### Changed
 - change handling of schema id uri
 
+## [0.3.4] - 2020-08-09
+### Changed
+- Change handling of schema id URI
+
 ## [0.3.3] - 2020-08-09
 ### Changed
-- sanitise class names on code generation
+- Sanitise class names on code generation
 
 ## [0.3.2] - 2020-08-09
 ### Changed
-- fix non-alphanumeric directory names on code generation
+- Fix non-alphanumeric directory names on code generation
 
 ## [0.3.1] - 2020-08-09
 ### Changed